CDPCP-12516 - clean up unused functions prior implementing dead code check
